Remove CoroutineExceptionHandler from kafka consumer job - Implementation of current CoroutineExceptionHandler will keep the coroutine "active", but the consumer will ble closed upon exception (use {} clause)
resolves navikt/dagpenger#226
