+++ build.gradle.kts
@@ -46,7 +46,7 @@ configurations.all {
 }
 
 dependencies {
-    implementation(kotlin("stdlib"))
+    implementation(kotlin("stdlib-jdk8"))
 
     implementation(Dagpenger.Events)
